pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Atlas: Learning to Optimally Memorize the Context at Test Time - LLM Training Implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "accelerate>=0.20.0",
    "tensorboard>=2.13.0",
    "einops>=0.7.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
    "ty>=0.0.1a14",
]
